move NotebookNode to top-level
should common to all nbformats

(I didn't change the old nbformats to  use it, just in case)
